requirements.txt
absl-py==0.13.0
amqp==5.1.1
asgiref==3.5.2
astunparse==1.6.3
auto-labeling-pipeline==0.1.21
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.23.10
botocore==1.26.10
cachetools==4.2.2
celery==5.2.7
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.4
django-celery-results==2.2.0
django-cleanup==6.0.0
django-cors-headers==3.12.0
django-drf-filepond==0.4.1
django-filter==21.1
django-health-check==3.16.5
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.9
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-xml==2.0.0
doccano==1.7.0
drf-yasg==1.20.0
environs==9.5.0
et-xmlfile==1.1.0
filetype==1.0.13
flatbuffers==1.12
furl==2.1.3
gast==0.3.3
google-api-core==2.8.1
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.0
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.2.4
lml==0.1.0
Markdown==3.3.4
MarkupSafe==2.1.1
marshmallow==3.16.0
matplotlib==3.4.3
natsort==8.1.0
nlpcloud==1.0.26
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.3
pandas==1.4.2
Pillow==8.3.1
prompt-toolkit==3.0.29
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.1
pydot==1.4.2
pydotplus==2.0.2
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-xlsx==0.6.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
scikit-learn==0.24.2
scipy==1.4.1
seqeval==1.2.2
shortuuid==1.0.9
six==1.15.0
sklearn==0.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.13.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==2.2.0
tqdm==4.64.0
typeguard==2.12.1
typing-extensions==3.7.4.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.6
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
Werkzeug==2.0.1
whitenoise==6.1.0
wincertstore==0.2
wrapt==1.12.1
